requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
pgvector
numpy
requests
httpx
python-multipart
google-cloud-aiplatform
